+++ src/sha/sha_point.c
@@ -2,6 +2,7 @@
 /* ------------------------------------------------------------------------- */
 #include <config.h>
 #include <stdio.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 #include <math.h>
@@ -350,7 +351,7 @@ void CHARM(sha_point)(const CHARM(point) *pnt, const REAL *f,
 
 
         /* ................................................................. */
-        int *ips  = NULL;
+        int64_t *ips  = NULL;
         REAL *ps  = NULL;
 #if !(CHARM_OPENMP)
         REAL *anm = NULL;
@@ -368,9 +369,9 @@ void CHARM(sha_point)(const CHARM(point) *pnt, const REAL *f,
         FFTWC(complex) *ftmp_out = NULL;
 
 
-        ips = (int *)CHARM(calloc_aligned)(SIMD_MEMALIGN,
-                                           SIMD_SIZE * SIMD_BLOCK * nmax,
-                                           sizeof(int));
+        ips = (int64_t *)CHARM(calloc_aligned)(SIMD_MEMALIGN,
+                                               SIMD_SIZE * SIMD_BLOCK * nmax,
+                                               sizeof(int64_t));
         if (ips == NULL)
         {
             FAILURE_glob = 1;
