+++ src/shs/shs_cell_kernel.c
@@ -2,6 +2,7 @@
 /* ------------------------------------------------------------------------- */
 #include <config.h>
 #include <stdio.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <math.h>
 #include "../leg/leg_func_xnum.h"
@@ -32,8 +33,8 @@ void CHARM(shs_cell_kernel)(unsigned long nmax,
                             REAL_SIMD u2,
                             const REAL *ps1,
                             const REAL *ps2,
-                            const int *ips1,
-                            const int *ips2,
+                            const int64_t *ips1,
+                            const int64_t *ips2,
                             REAL_SIMD *imm0,
                             REAL_SIMD *imm1,
                             REAL_SIMD *imm2,
