+++ src/shs/shs_cell_sctr.c
@@ -2,6 +2,7 @@
 /* ------------------------------------------------------------------------- */
 #include <config.h>
 #include <stdio.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <math.h>
 #include "../prec.h"
@@ -178,8 +179,8 @@ shared(ncell, FAILURE_glob, mur, err, rref)
         int FAILURE_priv = 0;
 
 
-        int  *ips1    = NULL;
-        int  *ips2    = NULL;
+        int64_t *ips1 = NULL;
+        int64_t *ips2 = NULL;
         REAL *ps1     = NULL;
         REAL *ps2     = NULL;
         REAL *t1v     = NULL;
@@ -199,15 +200,17 @@ shared(ncell, FAILURE_glob, mur, err, rref)
         REAL *cell_rv = NULL;
 
 
-        ips1 = (int *)CHARM(calloc_aligned)(SIMD_MEMALIGN, nmax * SIMD_SIZE,
-                                            sizeof(int));
+        ips1 = (int64_t *)CHARM(calloc_aligned)(SIMD_MEMALIGN,
+                                                nmax * SIMD_SIZE,
+                                                sizeof(int64_t));
         if (ips1 == NULL)
         {
             FAILURE_priv = 1;
             goto FAILURE_1_parallel;
         }
-        ips2 = (int *)CHARM(calloc_aligned)(SIMD_MEMALIGN, nmax * SIMD_SIZE,
-                                            sizeof(int));
+        ips2 = (int64_t *)CHARM(calloc_aligned)(SIMD_MEMALIGN,
+                                                nmax * SIMD_SIZE,
+                                                sizeof(int64_t));
         if (ips2 == NULL)
         {
             FAILURE_priv = 1;
