+++ src/shs/shs_point_sctr.c
@@ -2,6 +2,7 @@
 /* ------------------------------------------------------------------------- */
 #include <config.h>
 #include <stdio.h>
+#include <stdint.h>
 #include <stdlib.h>
 #include <math.h>
 #include "../prec.h"
@@ -177,7 +178,7 @@ shared(r_eq_rref, dr, dlat, dlon, npar, grad, dorder)
         int FAILURE_priv = 0;
 
 
-        int  *ips     = NULL;
+        int64_t *ips  = NULL;
         REAL *ps      = NULL;
         REAL *tv      = NULL;
         REAL *uv      = NULL;
@@ -191,9 +192,9 @@ shared(r_eq_rref, dr, dlat, dlon, npar, grad, dorder)
         REAL *enm     = NULL;
 
 
-        ips = (int *)CHARM(calloc_aligned)(SIMD_MEMALIGN,
-                                           nmax * SIMD_SIZE * SIMD_BLOCK,
-                                           sizeof(int));
+        ips = (int64_t *)CHARM(calloc_aligned)(SIMD_MEMALIGN,
+                                               nmax * SIMD_SIZE * SIMD_BLOCK,
+                                               sizeof(int64_t));
         if (ips == NULL)
         {
             FAILURE_priv = 1;
